requirements.txt
aniso8601==8.0.0
APScheduler==2.1.2
asn1crypto==0.24.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-RESTful==0.3.7
html5lib==1.0.1
httpcache==0.1.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
paramiko==2.6.0
Pillow==6.1.0
pycparser==2.19
pymal==0.6
PyNaCl==1.3.0
pysftp==0.2.9
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
schedule==0.6.0
singleton-factory==1.0
singleton3==1.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.5
webencodings==0.5.1
Werkzeug==0.16.0
